Change tracking-replacement to accept string and regex
The tracking-replacement was modified to take a string as a parameter as
well as regex. Previously, when passing a string to the replacer that
contained regex characters, the removeAll function recognized it as a
regex and incorrectly matched it. This has now been changed to check if
the input is a string or not, if it is then we mock the matched result
and continue with the replacement function.